refactor: Moved CV1 <-> CV2 conversion traits to traits.rs.
